Restructure code by putting them in different files according to specification section
